+++ src/registrar/views/utility/__init__.py
@@ -2,8 +2,6 @@ from .steps_helper import StepsHelper
 from .always_404 import always_404
 
 from .permission_views import (
-    DomainPermissionView,
     PortfolioMembersPermission,
-    DomainInvitationPermissionCancelView,
 )
 from .api_views import get_senior_official_from_federal_agency_json
