+++ src/registrar/views/utility/mixins.py
@@ -1,14 +1,6 @@
 """Permissions-related mixin classes."""
 
 from django.contrib.auth.mixins import PermissionRequiredMixin
-
-from registrar.models import (
-    Domain,
-    DomainRequest,
-    DomainInvitation,
-    DomainInformation,
-    UserDomainRole,
-)
 import logging
 
 
@@ -193,151 +185,6 @@ class PortfolioReportsPermission(PermissionsLoginMixin):
             return False
 
         return self.request.user.is_org_user(self.request)
-
-
-class DomainPermission(PermissionsLoginMixin):
-    """Permission mixin that redirects to domain if user has access,
-    otherwise 403"""
-
-    def has_permission(self):
-        """Check if this user has access to this domain.
-
-        The user is in self.request.user and the domain needs to be looked
-        up from the domain's primary key in self.kwargs["domain_pk"]
-        """
-        pk = self.kwargs["domain_pk"]
-
-        # test if domain in editable state
-        if not self.in_editable_state(pk):
-            return False
-
-        # if we need to check more about the nature of role, do it here.
-        return True
-
-    def in_editable_state(self, pk):
-        """Is the domain in an editable state"""
-
-        requested_domain = None
-        if Domain.objects.filter(id=pk).exists():
-            requested_domain = Domain.objects.get(id=pk)
-
-        # if domain is editable return true
-        if requested_domain and requested_domain.is_editable():
-            return True
-        return False
-
-    def can_access_other_user_domains(self, pk):
-        """Checks to see if an authorized user (staff or superuser)
-        can access a domain that they did not create or was invited to.
-        """
-
-        # Check if the user is permissioned...
-        user_is_analyst_or_superuser = self.request.user.has_perm(
-            "registrar.analyst_access_permission"
-        ) or self.request.user.has_perm("registrar.full_access_permission")
-
-        if not user_is_analyst_or_superuser:
-            return False
-
-        # Check if the user is attempting a valid edit action.
-        # In other words, if the analyst/admin did not click
-        # the 'Manage Domain' button in /admin,
-        # then they cannot access this page.
-        session = self.request.session
-        can_do_action = (
-            "analyst_action" in session
-            and "analyst_action_location" in session
-            and session["analyst_action_location"] == pk
-        )
-
-        if not can_do_action:
-            return False
-
-        # Analysts may manage domains, when they are in these statuses:
-        valid_domain_statuses = [
-            DomainRequest.DomainRequestStatus.APPROVED,
-            DomainRequest.DomainRequestStatus.IN_REVIEW,
-            DomainRequest.DomainRequestStatus.REJECTED,
-            DomainRequest.DomainRequestStatus.ACTION_NEEDED,
-            # Edge case - some domains do not have
-            # a status or DomainInformation... aka a status of 'None'.
-            # It is necessary to access those to correct errors.
-            None,
-        ]
-
-        requested_domain = None
-        if DomainInformation.objects.filter(id=pk).exists():
-            requested_domain = DomainInformation.objects.get(id=pk)
-
-        # if no domain information or domain request exist, the user
-        # should be able to manage the domain; however, if domain information
-        # and domain request exist, and domain request is not in valid status,
-        # user should not be able to manage domain
-        if (
-            requested_domain
-            and requested_domain.domain_request
-            and requested_domain.domain_request.status not in valid_domain_statuses
-        ):
-            return False
-
-        # Valid session keys exist,
-        # the user is permissioned,
-        # and it is in a valid status
-        return True
-
-
-class UserDeleteDomainRolePermission(PermissionsLoginMixin):
-    """Permission mixin for UserDomainRole if user
-    has access, otherwise 403"""
-
-    def has_permission(self):
-        """Check if this user has access to this domain request.
-
-        The user is in self.request.user and the domain needs to be looked
-        up from the domain's primary key in self.kwargs["pk"]
-        """
-        domain_pk = self.kwargs["pk"]
-        user_pk = self.kwargs["user_pk"]
-
-        if not self.request.user.is_authenticated:
-            return False
-
-        # Check if the UserDomainRole object exists, then check
-        # if the user requesting the delete has permissions to do so
-        has_delete_permission = UserDomainRole.objects.filter(
-            user=user_pk,
-            domain=domain_pk,
-            domain__permissions__user=self.request.user,
-        ).exists()
-
-        user_is_analyst_or_superuser = self.request.user.has_perm(
-            "registrar.analyst_access_permission"
-        ) or self.request.user.has_perm("registrar.full_access_permission")
-
-        if not (has_delete_permission or user_is_analyst_or_superuser):
-            return False
-
-        return True
-
-
-class DomainInvitationPermission(PermissionsLoginMixin):
-    """Permission mixin that redirects to domain invitation if user has
-    access, otherwise 403"
-
-    A user has access to a domain invitation if they have a role on the
-    associated domain.
-    """
-
-    def has_permission(self):
-        """Check if this user has a role on the domain of this invitation."""
-        if not self.request.user.is_authenticated:
-            return False
-
-        if not DomainInvitation.objects.filter(
-            id=self.kwargs["pk"], domain__permissions__user=self.request.user
-        ).exists():
-            return False
-        return True
 
 
 class UserProfilePermission(PermissionsLoginMixin):
